Added graph in T_Tracer to observe UL-SNR calculated using SRS

+++ openair1/PHY/NR_ESTIMATION/nr_ul_estimation.h
@@ -89,7 +89,7 @@ int nr_srs_channel_estimation(
                                     [gNB->frame_parms.ofdm_symbol_size * (1 << srs_pdu->num_symbols)],
     c16_t srs_estimated_channel_time[][1 << srs_pdu->num_ant_ports][gNB->frame_parms.ofdm_symbol_size],
     c16_t srs_estimated_channel_time_shifted[][1 << srs_pdu->num_ant_ports][gNB->frame_parms.ofdm_symbol_size],
-    int8_t *snr_per_rb,
+    int16_t *snr_per_rb,
     int8_t *snr);
 
 void nr_freq_equalization(NR_DL_FRAME_PARMS *frame_parms,
